Update.
Updated class Lot and service. Add controller for lk. Changed lk.html
(add names for field, output info of lots on the page). Updated class
Auction and service.

+++ src/main/java/com/RGR/Auction/repositories/LotRepositories.java
@@ -4,12 +4,14 @@ import java.util.List;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import com.RGR.Auction.models.CategoryModel;
+import com.RGR.Auction.models.Data;
 import com.RGR.Auction.models.Lot;
 
 @Repository
-public interface LotRepositories extends CrudRepository<Lot, Integer> {
+public interface LotRepositories extends CrudRepository<Lot, Long> {
 	
-    Lot findById(Long id);
+    Lot findById(long id);
     List <Lot> findAllByOrderByIdDesc();
-  
+	
 }
